Move dataclasses and enums to registry and types files
