+++ src/idi_company_info/processors/orchestrator.py
@@ -16,19 +16,12 @@ import pathlib
 import sys
 from dataclasses import dataclass
 from datetime import datetime
-from enum import Enum, StrEnum
-from typing import Any
+from enum import Enum
 
 from idi_company_info.common.logs import get_logger
-from idi_company_info.processors.identifier import (
-    IdentifierPipeline,
-    FilePaths,
-    BatchConfig,
-    ApiCredentials,
-    QueryType,
-)
-from idi_company_info.processors.IdentifierCik import IdentifierCik
-from idi_company_info.processors.IdentifierCusip import IdentifierCusip
+from idi_company_info.processors.identifier import IdentifierPipeline
+from idi_company_info.processors.types import ApiCredentials, BatchConfig, FilePaths, QueryType
+from idi_company_info.processors.registry import IDENTIFIER_REGISTRY, IdentifierSpec, IdentifierType
 
 
 # ---------------------------------------------------------------------------
@@ -42,64 +35,6 @@ class StageStatus(Enum):
     SUCCESS = "success"
     FAILED = "failed"
     SKIPPED = "skipped"
-
-
-# ---------------------------------------------------------------------------
-# Identifier type registry
-# ---------------------------------------------------------------------------
-
-class IdentifierType(StrEnum):
-    """Supported identifier types."""
-    CIK = "cik"
-    CIK_MATCH = "cik-match"
-    CUSIP = "cusip"
-    TICKER = "ticker"
-
-
-@dataclass
-class IdentifierSpec:
-    """Specification for a single identifier type.
-
-    Adding a new type requires only a new entry in IDENTIFIER_REGISTRY —
-    no other code needs to change.
-    """
-    cls: type[IdentifierPipeline]
-    query_type: QueryType
-    permid_filename: str
-    result_filename: str
-    failure_filename: str
-
-
-IDENTIFIER_REGISTRY: dict[IdentifierType, IdentifierSpec] = {
-    IdentifierType.CIK: IdentifierSpec(
-        cls=IdentifierCik,
-        query_type=QueryType.ENTITY_SEARCH,
-        permid_filename="permid_tracking_cik.json",
-        result_filename="company_info_cik.json",
-        failure_filename="failures_cik.json",
-    ),
-    IdentifierType.CUSIP: IdentifierSpec(
-        cls=IdentifierCusip,
-        query_type=QueryType.ENTITY_SEARCH,
-        permid_filename="permid_tracking_cusip.json",
-        result_filename="company_info_cusip.json",
-        failure_filename="failures_cusip.json",
-    ),
-    IdentifierType.CIK_MATCH: IdentifierSpec(
-        cls=IdentifierCik,
-        query_type=QueryType.RECORD_MATCH,
-        permid_filename="permid_tracking_cik_match.json",
-        result_filename="company_info_cik_match.json",
-        failure_filename="failures_cik_match.json",
-    ),
-    IdentifierType.TICKER: IdentifierSpec(
-        cls=IdentifierCusip,
-        query_type=QueryType.RECORD_MATCH,
-        permid_filename="permid_tracking_ticker.json",
-        result_filename="company_info_ticker.json",
-        failure_filename="failures_ticker.json",
-    ),
-}
 
 
 # ---------------------------------------------------------------------------
